Display Dropped Out reason in status card.
