+++ frontend/src/components/StudentHistory/Timeline.vue
@@ -9,7 +9,7 @@
                 <!-- If a student has dropped out, white dot is rendered -->
                 <div class="whitedot" v-if="item.description == 'DROPPED OUT'"></div>
                 <div class ="dot" v-else></div>
-                <div id="description"> {{item.description}}. {{item.reason}}.</div>
+                <div id="description"> {{item.description}}. {{item.reason}}</div>
         </div>
     </div>
