+++ frontend/src/store/students.js
@@ -12,8 +12,6 @@ export const studentActions = {
         commit(MUTATIONS.SET_STUDENTS, studentData)        
     },
 
-
-
     async createStudent({dispatch}, studentData){
         const payload = {
         method: "POST",
@@ -27,12 +25,84 @@ export const studentActions = {
         dispatch("getAllStudents")
         return response     
     },
-} 
 
+    // Update student status
+    async updateStudentStatus({ commit, dispatch }, { studentID, previousStatusString, nextStatusString, updatedBy, reason }) {
+        // PATCH student
+        const studentRequestOptions = {
+            method: "PATCH",
+            headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({StudentID: studentID, StatusString: nextStatusString})
+        }
 
+        // POST statusUpdate
+        const statusUpdateRequestOptions = {
+            method: "POST",
+            headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({
+            StudentID: studentID,
+            PreviousStatusString: previousStatusString,
+            NextStatusString: nextStatusString,
+            UpdatedBy: updatedBy,
+            ReasonString: reason
+            })
+        }
 
+        fetch("/api/statusUpdates", statusUpdateRequestOptions)
+        .then((response) => {
+            if(response.status !== 200) {
+                throw new Error(response);
+            }
+        })
+        .then(() => {
+            fetch("/api/students/" + studentID, studentRequestOptions)
+            .then((response) => {
+            if(response.status !== 200) {
+                throw new Error(response);
+            }
+            })
+        })
+        .then(() => {
+            dispatch('getAllStudents')
+        })
+        .catch((err) => {
+            console.error(err);
+        });
+    },
 
+    // Update student English proficiency
+    async updateStudentEnglishProficiency({ commit, dispatch }, { studentID, englishProficiency }) {
 
+    if (!englishProficiency) return;
+
+    // PATCH student
+    const studentRequestOptions = {
+        method: "PATCH",
+        headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({StudentID: studentID, EnglishProficiency: englishProficiency})
+    }
+
+    fetch("/api/students/" + studentID, studentRequestOptions)
+        .then(response => {
+            // If PATCH fails, return
+            if(response.status !== 200) {
+            console.log(response);
+            return;
+            }
+            dispatch('getAllStudents')
+        }
+        )
+    },
+}
 
 export const studentMutations = {
     [MUTATIONS.SET_STUDENTS](state, students){
